requirements.txt
alembic==1.0.7
arrow==0.12.1
backports.functools-lru-cache==1.5
beautifulsoup4==4.7.1
certifi==2018.11.29
chardet==3.0.4
cheroot==6.5.4
CherryPy==17.3.0
Click==7.0
contextlib2==0.5.5
feedparser==5.2.1
Flask==1.0.2
Flask-Migrate==2.3.1
Flask-SQLAlchemy==2.3.2
future==0.17.1
idna==2.7
itsdangerous==1.1.0
jaraco.functools==2.0
Jinja2==2.10
lxml==4.3.0
Mako==1.0.7
MarkupSafe==1.1.0
mintapi==1.35
mock==2.0.0
more-itertools==5.0.0
nokia==1.1.0
numpy==1.16.0
oauthlib==3.0.0
pandas==0.23.4
pbr==5.1.1
portend==2.3
pyarks==0.1.3
python-crontab==2.3.6
python-dateutil==2.7.5
python-editor==1.0.3
pytz==2018.9
requests==2.19.1
requests-file==1.4.3
requests-oauth==0.4.1
requests-oauthlib==1.0.0
selenium==3.141.0
selenium-requests==1.3
six==1.12.0
soupsieve==1.7.3
SQLAlchemy==1.2.16
SQLAlchemy-Utils==0.33.11
tempora==1.14
tldextract==2.2.0
urllib3==1.23
Werkzeug==0.14.1
xmltodict==0.11.0
zc.lockfile==1.4
